requirements.txt
instabot==0.108.0
requests==2.23.0
selenium==3.141.0
pillow>=6.2.1
beautifulsoup4